client: Rename ContainerServer to InstanceServer
And same to associated structs, alias old names.

Moving forward we will be using the new names in our own code and the
new Instance related functions will be using new versions of the various
argument structs.
