+++ client/interfaces.go
@@ -57,8 +57,8 @@ type ImageServer interface {
 	GetImageAlias(name string) (alias *api.ImageAliasesEntry, ETag string, err error)
 }
 
-// The ContainerServer type represents a full featured LXD server.
-type ContainerServer interface {
+// The InstanceServer type represents a full featured LXD server.
+type InstanceServer interface {
 	ImageServer
 
 	// Server functions
@@ -68,8 +68,8 @@ type ContainerServer interface {
 	HasExtension(extension string) (exists bool)
 	RequireAuthenticated(authenticated bool)
 	IsClustered() (clustered bool)
-	UseTarget(name string) (client ContainerServer)
-	UseProject(name string) (client ContainerServer)
+	UseTarget(name string) (client InstanceServer)
+	UseProject(name string) (client InstanceServer)
 
 	// Certificate functions
 	GetCertificateFingerprints() (fingerprints []string, err error)
@@ -86,7 +86,7 @@ type ContainerServer interface {
 	GetContainer(name string) (container *api.Container, ETag string, err error)
 	CreateContainer(container api.ContainersPost) (op Operation, err error)
 	CreateContainerFromImage(source ImageServer, image api.Image, imgcontainer api.ContainersPost) (op RemoteOperation, err error)
-	CopyContainer(source ContainerServer, container api.Container, args *ContainerCopyArgs) (op RemoteOperation, err error)
+	CopyContainer(source InstanceServer, container api.Container, args *ContainerCopyArgs) (op RemoteOperation, err error)
 	UpdateContainer(name string, container api.ContainerPut, ETag string) (op Operation, err error)
 	RenameContainer(name string, container api.ContainerPost) (op Operation, err error)
 	MigrateContainer(name string, container api.ContainerPost) (op Operation, err error)
@@ -105,7 +105,7 @@ type ContainerServer interface {
 	GetContainerSnapshots(containerName string) (snapshots []api.ContainerSnapshot, err error)
 	GetContainerSnapshot(containerName string, name string) (snapshot *api.ContainerSnapshot, ETag string, err error)
 	CreateContainerSnapshot(containerName string, snapshot api.ContainerSnapshotsPost) (op Operation, err error)
-	CopyContainerSnapshot(source ContainerServer, containerName string, snapshot api.ContainerSnapshot, args *ContainerSnapshotCopyArgs) (op RemoteOperation, err error)
+	CopyContainerSnapshot(source InstanceServer, containerName string, snapshot api.ContainerSnapshot, args *ContainerSnapshotCopyArgs) (op RemoteOperation, err error)
 	RenameContainerSnapshot(containerName string, name string, container api.ContainerSnapshotPost) (op Operation, err error)
 	MigrateContainerSnapshot(containerName string, name string, container api.ContainerSnapshotPost) (op Operation, err error)
 	DeleteContainerSnapshot(containerName string, name string) (op Operation, err error)
@@ -205,8 +205,8 @@ type ContainerServer interface {
 	UpdateStoragePoolVolume(pool string, volType string, name string, volume api.StorageVolumePut, ETag string) (err error)
 	DeleteStoragePoolVolume(pool string, volType string, name string) (err error)
 	RenameStoragePoolVolume(pool string, volType string, name string, volume api.StorageVolumePost) (err error)
-	CopyStoragePoolVolume(pool string, source ContainerServer, sourcePool string, volume api.StorageVolume, args *StoragePoolVolumeCopyArgs) (op RemoteOperation, err error)
-	MoveStoragePoolVolume(pool string, source ContainerServer, sourcePool string, volume api.StorageVolume, args *StoragePoolVolumeMoveArgs) (op RemoteOperation, err error)
+	CopyStoragePoolVolume(pool string, source InstanceServer, sourcePool string, volume api.StorageVolume, args *StoragePoolVolumeCopyArgs) (op RemoteOperation, err error)
+	MoveStoragePoolVolume(pool string, source InstanceServer, sourcePool string, volume api.StorageVolume, args *StoragePoolVolumeMoveArgs) (op RemoteOperation, err error)
 	MigrateStoragePoolVolume(pool string, volume api.StorageVolumePost) (op Operation, err error)
 
 	// Storage volume snapshot functions ("storage_api_volume_snapshots" API extension)
@@ -241,15 +241,6 @@ type ConnectionInfo struct {
 	URL         string
 	SocketPath  string
 	Project     string
-}
-
-// The ContainerBackupArgs struct is used when creating a container from a backup
-type ContainerBackupArgs struct {
-	// The backup file
-	BackupFile io.Reader
-
-	// Storage pool to use
-	PoolName string
 }
 
 // The BackupFileRequest struct is used for a backup download request
@@ -354,111 +345,4 @@ type StoragePoolVolumeCopyArgs struct {
 // during storage volume move
 type StoragePoolVolumeMoveArgs struct {
 	StoragePoolVolumeCopyArgs
-}
-
-// The ContainerCopyArgs struct is used to pass additional options during container copy
-type ContainerCopyArgs struct {
-	// If set, the container will be renamed on copy
-	Name string
-
-	// If set, the container running state will be transferred (live migration)
-	Live bool
-
-	// If set, only the container will copied, its snapshots won't
-	ContainerOnly bool
-
-	// The transfer mode, can be "pull" (default), "push" or "relay"
-	Mode string
-
-	// API extension: container_incremental_copy
-	// Perform an incremental copy
-	Refresh bool
-}
-
-// The ContainerSnapshotCopyArgs struct is used to pass additional options during container copy
-type ContainerSnapshotCopyArgs struct {
-	// If set, the container will be renamed on copy
-	Name string
-
-	// The transfer mode, can be "pull" (default), "push" or "relay"
-	Mode string
-
-	// API extension: container_snapshot_stateful_migration
-	// If set, the container running state will be transferred (live migration)
-	Live bool
-}
-
-// The ContainerConsoleArgs struct is used to pass additional options during a
-// container console session
-type ContainerConsoleArgs struct {
-	// Bidirectional fd to pass to the container
-	Terminal io.ReadWriteCloser
-
-	// Control message handler (window resize)
-	Control func(conn *websocket.Conn)
-
-	// Closing this Channel causes a disconnect from the container's console
-	ConsoleDisconnect chan bool
-}
-
-// The ContainerConsoleLogArgs struct is used to pass additional options during a
-// container console log request
-type ContainerConsoleLogArgs struct {
-}
-
-// The ContainerExecArgs struct is used to pass additional options during container exec
-type ContainerExecArgs struct {
-	// Standard input
-	Stdin io.ReadCloser
-
-	// Standard output
-	Stdout io.WriteCloser
-
-	// Standard error
-	Stderr io.WriteCloser
-
-	// Control message handler (window resize, signals, ...)
-	Control func(conn *websocket.Conn)
-
-	// Channel that will be closed when all data operations are done
-	DataDone chan bool
-}
-
-// The ContainerFileArgs struct is used to pass the various options for a container file upload
-type ContainerFileArgs struct {
-	// File content
-	Content io.ReadSeeker
-
-	// User id that owns the file
-	UID int64
-
-	// Group id that owns the file
-	GID int64
-
-	// File permissions
-	Mode int
-
-	// File type (file or directory)
-	Type string
-
-	// File write mode (overwrite or append)
-	WriteMode string
-}
-
-// The ContainerFileResponse struct is used as part of the response for a container file download
-type ContainerFileResponse struct {
-	// User id that owns the file
-	UID int64
-
-	// Group id that owns the file
-	GID int64
-
-	// File permissions
-	Mode int
-
-	// File type (file or directory)
-	Type string
-
-	// If a directory, the list of files inside it
-	Entries []string
 }
